pip_requirements.txt
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
bottle==0.12.18
certifi==2020.4.5.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
entrypoints==0.3
igl==2.2.0
importlib-metadata==1.6.1
instaviz==0.6.0
ipydatawidgets==4.0.1
ipykernel==5.3.0
ipyscales==0.4.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
joblib==0.15.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
meshplot==0.3.3
mistune==0.8.4
more-itertools==8.2.0
nbconvert==5.6.1
nbformat==5.0.6
nose==1.3.7
notebook==6.0.3
numpy==1.18.5
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
py==1.8.1
Pygments==2.6.1
pylint==2.4.4
pymesh2==0.2.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.1
python-dateutil==2.8.1
pythreejs==2.2.0
pytz==2020.1
pyzmq==19.0.1
scikit-learn==0.23.1
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==4.3.3
traittypes==0.2.1
typed-ast==1.4.1
vedo==2020.3.2
vtk==8.1.2
wcwidth==0.2.4
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
